import adjustments without move
