--- cstar/base/__init__.py
@@ -1,14 +1,0 @@
-from cstar.base import exceptions
-from cstar.base.additional_code import AdditionalCode
-from cstar.base.discretization import Discretization
-from cstar.base.external_codebase import ExternalCodeBase
-from cstar.base.input_dataset import InputDataset
-
-__all__ = [
-    "Component",
-    "ExternalCodeBase",
-    "AdditionalCode",
-    "InputDataset",
-    "Discretization",
-    "exceptions",
-]
