+++ cstar/cli/workplan/compose.py
@@ -6,10 +6,9 @@ from pathlib import Path
 
 import typer
 
+from cstar.base.env import ENV_CSTAR_STATE_HOME, get_env_item
 from cstar.base.utils import (
-    ENV_CSTAR_STATE_HOME,
     additional_files_dir,
-    get_env_item,
 )
 from cstar.orchestration.dag_runner import build_and_run_dag
