--- cstar/marbl/__init__.py
@@ -1,5 +1,0 @@
-from cstar.marbl.external_codebase import MARBLExternalCodeBase
-
-__all__ = [
-    "MARBLExternalCodeBase",
-]
