+++ cstar/orchestration/orchestration.py
@@ -6,9 +6,10 @@ from pathlib import Path
 
 import networkx as nx
 
+from cstar.base.env import ENV_CSTAR_STATE_HOME
 from cstar.base.exceptions import CstarExpectationFailed
 from cstar.base.log import LoggingMixin
-from cstar.base.utils import ENV_CSTAR_STATE_HOME, slugify
+from cstar.base.utils import slugify
 from cstar.orchestration.models import Step, Workplan
 from cstar.orchestration.utils import (
     ENV_CSTAR_ORCH_RUNID,
